Moved Api methods in their respecting class
