+++ OsuPlayer.Network/API/ApiEndpoints/ApiUserEndpoint.cs
@@ -164,32 +164,4 @@ public partial class ApiAsync
 
         await ApiRequestAsync<UserOnlineStatusModel>("users", "setUserOnlineStatus", data);
     }
-
-    /// <summary>
-    /// Gets API statistics from /api/statistics/getApiStatisticsV2
-    /// </summary>
-    /// <returns>Statistics as an <see cref="ApiStatisticsV3Model" /></returns>
-    public static async Task<ApiStatisticsV3Model> GetApiStatistics()
-    {
-        // default is new value due to possible null
-        return await GetRequestAsync<ApiStatisticsV3Model>("statistics", "getApiStatisticsV2") ?? new ApiStatisticsV3Model();
-    }
-
-    /// <summary>
-    /// Gets beatmap count from /api/beatmaps/getBeatmapsCount
-    /// </summary>
-    /// <returns>Beatmap count as an integer</returns>
-    public static async Task<int> GetBeatmapsCount()
-    {
-        return await GetRequestAsync<int>("beatmaps", "getBeatmapsCount");
-    }
-
-    /// <summary>
-    /// Gets the amount of storage used on the database
-    /// </summary>
-    /// <returns>Disk space used (in MB) as a float</returns>
-    public static async Task<float> GetStorageAmount()
-    {
-        return await GetRequestAsync<float>("osuplayer", "getStorageAmount");
-    }
 }
